Integrate `get_loss` and `decode` into `Model`
