+++ parser/cmds/cmd.py
@@ -2,11 +2,10 @@
 
 import os
 from parser.utils import Embedding
-from parser.utils.alg import eisner
 from parser.utils.common import bos, pad, unk
 from parser.utils.corpus import CoNLL, Corpus
 from parser.utils.field import BertField, CharField, Field
-from parser.utils.fn import ispunct, istree, numericalize
+from parser.utils.fn import ispunct, numericalize
 from parser.utils.metric import AttachmentMetric
 
 import torch
@@ -60,7 +59,6 @@ class CMD(object):
             self.ARC, self.REL = self.fields.HEAD, self.fields.DEPREL
         self.puncts = torch.tensor([i for s, i in self.WORD.vocab.stoi.items()
                                     if ispunct(s)]).to(args.device)
-        self.criterion = nn.CrossEntropyLoss()
 
         args.update({
             'n_words': self.WORD.vocab.n_init,
@@ -86,14 +84,14 @@ class CMD(object):
             # ignore the first token of each sentence
             mask[:, 0] = 0
             s_arc, s_rel = self.model(words, feats)
-            loss = self.get_loss(s_arc, s_rel, arcs, rels, mask)
+            loss = self.model.get_loss(s_arc, s_rel, arcs, rels, mask)
             loss.backward()
             nn.utils.clip_grad_norm_(self.model.parameters(),
                                      self.args.clip)
             self.optimizer.step()
             self.scheduler.step()
 
-            arc_preds, rel_preds = self.decode(s_arc, s_rel, mask)
+            arc_preds, rel_preds = self.model.decode(s_arc, s_rel, mask)
             # ignore all punctuation if not specified
             if not self.args.punct:
                 mask &= words.unsqueeze(-1).ne(self.puncts).all(-1)
@@ -114,8 +112,8 @@ class CMD(object):
             # ignore the first token of each sentence
             mask[:, 0] = 0
             s_arc, s_rel = self.model(words, feats)
-            loss = self.get_loss(s_arc, s_rel, arcs, rels, mask)
-            arc_preds, rel_preds = self.decode(s_arc, s_rel, mask)
+            loss = self.model.get_loss(s_arc, s_rel, arcs, rels, mask)
+            arc_preds, rel_preds = self.model.decode(s_arc, s_rel, mask)
             # ignore all punctuation if not specified
             if not self.args.punct:
                 mask &= words.unsqueeze(-1).ne(self.puncts).all(-1)
@@ -136,7 +134,7 @@ class CMD(object):
             mask[:, 0] = 0
             lens = mask.sum(1).tolist()
             s_arc, s_rel = self.model(words, feats)
-            arc_preds, rel_preds = self.decode(s_arc, s_rel, mask)
+            arc_preds, rel_preds = self.model.decode(s_arc, s_rel, mask)
             all_arcs.extend(arc_preds[mask].split(lens))
             all_rels.extend(rel_preds[mask].split(lens))
             if self.args.prob:
@@ -147,27 +145,3 @@ class CMD(object):
         all_probs = [[round(p, 4) for p in seq.tolist()] for seq in all_probs]
 
         return all_arcs, all_rels, all_probs
-
-    def get_loss(self, s_arc, s_rel, arcs, rels, mask):
-        s_arc, arcs = s_arc[mask], arcs[mask]
-        s_rel, rels = s_rel[mask], rels[mask]
-        s_rel = s_rel[torch.arange(len(arcs)), arcs]
-        arc_loss = self.criterion(s_arc, arcs)
-        rel_loss = self.criterion(s_rel, rels)
-        loss = arc_loss + rel_loss
-
-        return loss
-
-    def decode(self, s_arc, s_rel, mask):
-        lens = mask.sum(1)
-        # prevent self-loops
-        s_arc.diagonal(0, 1, 2).fill_(float('-inf'))
-        arc_preds = s_arc.argmax(-1)
-        bad = [not istree(sequence[:l+1], self.args.proj)
-               for l, sequence in zip(lens.tolist(), arc_preds.tolist())]
-        if self.args.tree and any(bad):
-            arc_preds[bad] = eisner(s_arc[bad], mask[bad])
-        rel_preds = s_rel.argmax(-1)
-        rel_preds = rel_preds.gather(-1, arc_preds.unsqueeze(-1)).squeeze(-1)
-
-        return arc_preds, rel_preds
